Improved compiler flags settings and updated DRM files
Description:
- Improved compiler flags settings,
- Updated DRM files,
- Updated device ids for PVC and ACM,
- Max value definition added to SAMPLER_L3_READ on ACM, MTL,
- Added SAMPLER_L3_BYTE_READ metrics to ACM, MTL,
- Updated short names for MTL,
- Added XVE_COMPUTE_THREAD_COUNT for MTL,
- Fixed Coverity issues.

+++ instrumentation/metrics_discovery/common/internal/md_adapter_group.cpp
@@ -441,7 +441,7 @@ namespace MetricsDiscoveryInternal
         MD_CHECK_PTR_RET( adapter, CC_ERROR_NO_MEMORY );
 
         m_adapterVector.push_back( adapter );
-        m_params.AdapterCount = m_adapterVector.size();
+        m_params.AdapterCount = static_cast<uint32_t>( m_adapterVector.size() );
 
         const TAdapterParamsLatest* adapterParams = adapter->GetParams();
         const uint32_t              adapterId     = adapter->GetAdapterId();
@@ -450,6 +450,7 @@ namespace MetricsDiscoveryInternal
         MD_LOG_A( adapterId, LOG_INFO, "Platform ID: %u", adapterParams->Platform );
         MD_LOG_A( adapterId, LOG_INFO, "Device ID: %x", adapterParams->DeviceId );
         MD_LOG_A( adapterId, LOG_INFO, "Adapter ID: %u", adapterId );
+
         return CC_OK;
     }
